Added author, upvote, and dropdown controls to DiscussionDrawer
